Refactor by review comments
