+++ MediaGalleryUi/etc/adminhtml/di.xml
@@ -6,7 +6,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="Magento\MediaGalleryUi\Controller\Adminhtml\Directories\GetTree">
+    <type name="Magento\MediaGalleryUi\Model\Directories\FolderTree">
         <arguments>
             <argument name="path" xsi:type="string">media</argument>
         </arguments>
