+++ MediaGalleryUi/view/adminhtml/web/css/source/_module.less
@@ -23,12 +23,12 @@
             width: 20px;
         }
 
-        .jstree-default .jstree-no-dots .jstree-leaf > ins{ 
-            background-image: none !important; 
+        .jstree-default .jstree-no-dots .jstree-leaf > ins { 
+            background-image: none; 
         }
 
         .jstree-default ins {
-            background-image: url("../../images/d.png") !important;
+            background-image: url("@{baseDir}Magento_MediaGalleryUi/images/d.png");
         }
 
         .jstree a {
