+++ MediaGalleryUi/view/adminhtml/web/js/grid/columns/image.js
@@ -11,7 +11,6 @@ define([
         defaults: {
             bodyTmpl: 'Magento_MediaGalleryUi/grid/columns/image',
             selected: null,
-            getDirectoryTreeUrl: 'media_gallery/directories/gettree',
             fields: {
                 id: 'id',
                 url: 'url'
